rename 'QueryBuilder_Constants' to 'QueryBuilder_Const'
