+++ packages/legend-studio-preset-query-builder/src/models/protocols/pure/v1/V1_QueryBuilder_FunctionExpressionBuilder.ts
@@ -41,7 +41,7 @@ import {
   returnUndefOnError,
   UnsupportedOperationError,
 } from '@finos/legend-studio-shared';
-import { SUPPORTED_FUNCTIONS } from '../../../../QueryBuilder_Constants';
+import { SUPPORTED_FUNCTIONS } from '../../../../QueryBuilder_Const';
 
 const buildBaseSimpleFunctionExpression = (
   processedParameters: ValueSpecification[],
